requirements.txt
# Gen AI Platform - Requirements

# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1

# Google AI SDK
google-genai==0.2.2

# Image processing
Pillow>=10.0.0

# Database
SQLAlchemy>=2.0.35

# Async support
asyncio==3.4.3

# Environment variables
python-dotenv==1.0.0

# Utilities
pathlib==1.0.1

# Optional: Development tools
python-dateutil==2.8.2

# Optional: Production server
gunicorn==21.2.0

# Optional: Security
python-jose[cryptography]==3.3.0

# Optional: CORS support
flask-cors==4.0.0
